+++ modules/terraform-zscc-gwlbendpoint-aws/main.tf
@@ -16,6 +16,7 @@ data "aws_partition" "current" {}
 # if no explicit principals are configured in var.allowed_principals
 ################################################################################
 resource "aws_vpc_endpoint_service" "gwlb_vpce_service" {
+  count                      = var.byo_endpoint_service_name != null ? 0 : 1
   allowed_principals         = coalescelist(var.allowed_principals, ["arn:${data.aws_partition.current.partition}:iam::${data.aws_caller_identity.current.id}:root"])
   acceptance_required        = var.acceptance_required
   gateway_load_balancer_arns = [var.gwlb_arn]
@@ -25,15 +26,22 @@ resource "aws_vpc_endpoint_service" "gwlb_vpce_service" {
   )
 }
 
+# Or reference an existing Endpoint Service
+data "aws_vpc_endpoint_service" "gwlb_vpce_service_selected" {
+  count        = var.byo_endpoint_service_name != null ? 1 : 0
+  service_name = var.byo_endpoint_service_name
+}
+
+
 
 ################################################################################
 # Create the GWLB Endpoint ENIs per list of subnet IDs specified
 ################################################################################
 resource "aws_vpc_endpoint" "gwlb_vpce" {
   count             = length(var.subnet_ids)
-  service_name      = aws_vpc_endpoint_service.gwlb_vpce_service.service_name
+  service_name      = try(data.aws_vpc_endpoint_service.gwlb_vpce_service_selected[0].service_name, aws_vpc_endpoint_service.gwlb_vpce_service[0].service_name)
   subnet_ids        = [element(var.subnet_ids, count.index)]
-  vpc_endpoint_type = aws_vpc_endpoint_service.gwlb_vpce_service.service_type
+  vpc_endpoint_type = try(data.aws_vpc_endpoint_service.gwlb_vpce_service_selected[0].service_type, aws_vpc_endpoint_service.gwlb_vpce_service[0].service_type)
   vpc_id            = var.vpc_id
 
   tags = merge(var.global_tags,